LayoutTests/fullscreen/full-screen-iframe-with-max-width-height-expected.txt


EVENT(webkitfullscreenchange) TEST(testFrame.getBoundingClientRect().right == document.body.getBoundingClientRect().right && testFrame.getBoundingClientRect().bottom == document.body.getBoundingClientRect().bottom) OK
END OF TEST
